fix: calculateDday utils로 옮김
